Fixed for the PRB allocation issue assert in DLSCH scheduler
The following ASSERT is tirggered while testing 16 UE setup.
Corrected the get_start_stop_allocation calculation to fix the issue.
The implementation was wrong. If RBsize is 5 and RBstart is 0, then RBstop
is 4 not 5 as it was before this fix. This lead to an assertion
in multi-ue scenario.

Assertion (NPRB>0 && (NPRB + RBstart <= BWPsize)) failed!
In PRBalloc_to_locationandbandwidth0() /home/ran/common/utils/nr/nr_common.c:286
Illegal NPRB/RBstart Configuration (1,48) for BWPsize 48

Exiting execution