F1AP: message encoders should no trigger GTP
The F1AP message encoders and decoders should do only that, encoding and
decoding messages. For the actual logic handling, including set up and
teardown of bearers, we should use the handlers. In the present case of
the CU, this happens already in the CU-UP/E1 handlers in
cucp_cuup_handler.c.