Fixes after merging develop
Reversion of changes to tx_data.request and rx_data.indication on functions not related to Aerial L1.
Addition of ifdefs to include files related to Aerial integration
Fix DMRS ports id and numDmrsCdmGrpsNoData to only set changed values when on Aerial mode.
Fix dci_pdu->num_antenna_ports value to only set to 2 when on Aerial, in other cases, it remains with the previously used value.
Removal of commented define.